requirements.txt
dj-database-url==0.5.0
Django==3.1.0
django-s3direct==1.1.8
django-tinymce==3.0.2
django-crispy-forms==1.7.2
gunicorn==19.9.0
Jinja2==2.10.1
Pillow==6.2.1
psycopg2-binary==2.8.3
pyjade==4.0.0
requests==2.22.0
whitenoise==4.1.2